moved src into package
