+++ .gitignore
@@ -1,0 +1,3 @@
+lib/
+.idea/
+*.iml
